# api/http/v1/swagger.yaml
swagger: "2.0"
basePath: /api/v1
consumes:
  - application/json
info:
  version: 1.0.0
  title: title
  description: description
  termsOfService: terms
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
paths:
...
  /account/profiles/{profile_id}/checkups/{checkup_id}:
    get:
      operationId: Checkup
      tags:
        - Checkups
      parameters:
        - name: profile_id
          in: path
          type: string
          required: true
          description: profile id
        - name: checkup_id
          in: path
          type: string
          required: true
          description: checkup id
      produces:
        - application/json
      responses:
        200:
          description: Checkup details
          schema:
            $ref: '#/definitions/CheckupWithResults'
        404:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
    patch:
      operationId: UpdateCheckup
      tags:
        - Checkups
      parameters:
        - name: profile_id
          in: path
          type: string
          required: true
          description: profile id
        - name: checkup_id
          in: path
          type: string
          required: true
          description: checkup id
      produces:
        - application/json
      responses:
        200:
          description: Checkup details
          schema:
            $ref: '#/definitions/CheckupUpdateRequest'
        404:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
    delete:
      operationId: CheckupDelete
      tags:
        - Checkups
      parameters:
        - name: profile_id
          in: path
          type: string
          required: true
          description: profile id
        - name: checkup_id
          in: path
          type: string
          required: true
          description: checkup id
      produces:
        - application/json
      responses:
        200:
          description: Checkup details
          schema:
            $ref: '#/definitions/CheckupResult'
        404:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
...
  /account/profiles:
    get:
      operationId: Profiles
      produces:
        - application/json
      responses:
        200:
          description: List of account profiles
          schema:
            $ref: '#/definitions/CheckupsResponse'
        400:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        401:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        404:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        500:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
      tags:
        - Profiles
    post:
      operationId: CreateProfile
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          schema:
            $ref: '#/definitions/Person'
      responses:
        200:
          description: Create profile for account
          schema:
            $ref: '#/definitions/Success'
        201:
          description: Create profile for account
          schema:
            $ref: '#/definitions/Success'
        400:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        401:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        404:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        500:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
      tags:
        - Profiles
  /account/profiles/{profile_id}:
    get:
      operationId: Profile
      tags:
        - Profiles
      parameters:
        - name: profile_id
          in: path
          required: true
          type: string
      produces:
        - application/json
      responses:
        200:
          description: Account profile
          schema:
            $ref: '#/definitions/CheckupsResponse'
        400:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        401:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        404:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        500:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
    patch:
      operationId: UpdateProfile
      tags:
        - Profiles
      parameters:
        - name: profile_id
          in: path
          required: true
          type: string
      produces:
        - application/json
      responses:
        200:
          description: Update profile for account
          schema:
            $ref: '#/definitions/Success'
        201:
          description: Create profile for account
          schema:
            $ref: '#/definitions/Success'
        400:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        401:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        404:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        500:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
    delete:
      operationId: DeleteProfile
      tags:
        - Profiles
      parameters:
        - name: profile_id
          in: path
          required: true
          type: string
      produces:
        - application/json
      responses:
        200:
          description: Delete success status
          schema:
            $ref: '#/definitions/Success'
        400:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        401:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        404:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        500:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
...
  /marker:
    get:
      operationId: Markers

      produces:
        - application/json
      responses:
        200:
          description: List of available markers
          schema:
            $ref: '#/definitions/Chart'
        404:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
      tags:
        - Assessment
  /dictionaries:
    get:
      operationId: GetDictionaries
      summary: Get dictionaries data
      description: Returns lists of available markers and units
      tags:
        - Dictionaries
      produces:
        - application/json
      responses:
        200:
          description: Dictionary data with markers and units
          schema:
            $ref: '#/definitions/DictionariesResponse'
        404:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'
        500:
          description: Base ApiError
          schema:
            $ref: '#/definitions/ApiError'

# securityDefinitions:
#   ApiKeyAuth:
#     type: apiKey
#     in: header
#     name: Authorization
definitions:
...
  Chart:
    description: Data to represent on chart (time series)
    type: object
    properties:
      title:
        type: string
        example: "Custom dashboard"
      min_x:
        type: number
        example: 0
      min_y:
        type: number
        example: 0
      max_x:
        type: number
        example: 0
      max_y:
        type: number
        example: 0
      data:
        type: array
        items:
          $ref: '#/definitions/TimeSeriesData'
      annotations:
        type: array
        items:
          $ref: '#/definitions/ChartAnnotation'
#      timestamps:
#        type: array
#        items:
#          type: integer
...
  TimeSeriesData:
    description: Time series data point
    properties:
      id:
        type: integer
        x-omitempty: false
      title:
        type: string
        x-omitempty: false
        example: "hemoglobin"
      unit:
        type: string
        x-omitempty: false
        example: "g/dL"
      type:
        type: string
        x-omitempty: false
        example: "line"
      style:
        type: object
        $ref: '#/definitions/ChartStyle'
      spots:
        type: array
        items:
          $ref: '#/definitions/ChartSpot'
  ChartSpot:
    description: Chart point coordinates
    type: object
    properties:
      x:
        type: number
        x-omitempty: false
      y:
        type: number
        x-omitempty: false
  ChartAnnotation:
    description: Chart annotation
    type: object
    properties:
      start:
        type: number
        x-omitempty: false
      end:
        type: number
        x-omitempty: false
      axis:
        type: string
        x-omitempty: false
      style:
        type: object
        $ref: '#/definitions/ChartStyle'
  ChartStyle:
    description: Chart item style
    type: object
    properties:
      primary_color:
        type: integer
        x-omitempty: false
      secondary_color:
        type: integer
        x-omitempty: false
      opacity:
        type: number
        x-omitempty: false
...
  CheckupsResponse:
    description: list of checkups
    type: array
    items:
      $ref: '#/definitions/CheckupWithResults'
  CheckupWithResults:
    description: Checkup with results object
    type: object
    properties:
      checkup:
        $ref: '#/definitions/Checkup'
      results:
        type: array
        items:
          $ref: '#/definitions/Marker'
  Checkup:
    description: Checkup object
    type: object
    properties:
      id:
        type: integer
        x-omitempty: false
      title:
        type: string
        x-omitempty: false
      profile:
        $ref: '#/definitions/Profile'
      lab:
        $ref: '#/definitions/Lab'
      date:
        type: integer
        x-omitempty: false
      file_id:
        type: string
        x-omitempty: true
      status:
        type: string
        enum: [ unverified, verified ]
        x-omitempty: false
      material:
        type: string
        x-omitempty: true
      tags:
        type: array
        items:
          $ref: '#/definitions/Tag'
  CheckupResult:
    description: Recognized test results
    type: object
    properties:
      profile:
        $ref: '#/definitions/Profile'
      lab:
        $ref: '#/definitions/Lab'
      date:
        type: integer
        x-omitempty: false
      results:
        type: array
        items:
          $ref: '#/definitions/Marker'
  CheckupUpdateRequest:
    description: Update checkup data
    type: object
    properties:
      checkup:
        $ref: '#/definitions/Checkup'
      results_to_add:
        type: array
        items:
          $ref: '#/definitions/Marker'
      results_ids_to_delete:
        type: array
        items:
          type: integer
          format: int64
...
  Unit:
    description: marker result unit object
    type: object
    properties:
      id:
        type: number
        format: integer
        x-omitempty: false
      name:
        type: string
        x-omitempty: false
      full_name:
        type: string
        x-omitempty: true
      description:
        type: string
        x-omitempty: true
      unrecognized_name:
        type: string
        x-omitempty: true
  Tag:
    description: User tag object
    type: object
    properties:
      name:
        type: string
        x-omitempty: false
      id:
        type: number
        format: integer
        x-omitempty: false
      color:
        type: string
        x-omitempty: false
      icon:
        type: string
        x-omitempty: false
  Lab:
    description: Lab object
    type: object
    properties:
      name:
        type: string
        x-omitempty: false
      id:
        type: number
        format: integer
        x-omitempty: false
  Person:
    description: Person object
    type: object
    required:
      - first_name
      - gender
    properties:
      id:
        type: string
        format: string
      first_name:
        type: string
        x-omitempty: true
      date_of_birth:
        type: string
        #        format: date
        x-omitempty: true
      gender:
        type: string
        enum: [ M, F ]
        x-omitempty: true
      contacts:
        type: array
        x-omitempty: true
        items:
          $ref: '#/definitions/Contact'
...
  Profile:
    description: Profile object
    type: object
    required:
      - uuid
    properties:
      id:
        type: string
        format: string
      uuid:
        type: string
        format: string
      f_name:
        type: string
        x-omitempty: false
      l_name:
        type: string
        x-omitempty: false
      date_of_birth:
        type: string
        x-omitempty: false
      gender:
        type: string
        enum: [M, F]
      access:
        type: string
        enum: [ owner, editor, reader ]
      associated_user:
        type: string
        format: uuid
      contacts:
        type: array
        items:
          $ref: '#/definitions/Contact'
  Contact:
    description: Contact details
    type: object
    properties:
      value:
        type: string
      type:
        type: string
      is_verified:
        type: boolean
...
  Marker:
    description: Marker with value
    type: object
    properties:
      id:
        type: number
        format: integer
        x-omitempty: false
        description: "Marker ID from predefined list"
      result_id:
        type: number
        format: integer
        x-omitempty: true
        description: "Internal result ID"
      name:
        type: string
        x-omitempty: false
      unrecognized_name:
        type: string
        x-omitempty: true
      value:
        type: number
        x-omitempty: false
      date:
        type: integer
        x-omitempty: false
      reference_range:
        $ref: '#/definitions/ReferenceRange'
      unit:
        $ref: '#/definitions/Unit'
  ReferenceRange:
    description: Reference range for marker
    type: object
    properties:
      min:
        type: number
        x-omitempty: true
      max:
        type: number
        x-omitempty: true
...
  ApiError:
    description: Error response for api
    type: object
    properties:
      code:
        format: int64
        type: integer
        x-omitempty: false
      error:
        $ref: '#/definitions/ErrorMessage'
  ErrorMessage:
    description: Base error message
    type: object
    properties:
      data:
        type: object
      message:
        type: string
  Success:
    type: object
    properties:
      success:
        type: string
        x-omitempty: false
  DictionariesResponse:
    description: Dictionary data with markers and units
    type: object
    properties:
      markers:
        type: array
        items:
          $ref: '#/definitions/MarkerDefinition'
      units:
        type: array
        items:
          $ref: '#/definitions/Unit'
      labs:
        type: array
        items:
          $ref: '#/definitions/Lab'
  MarkerDefinition:
    description: Marker definition without value and date
    type: object
    properties:
      id:
        type: number
        format: integer
        x-omitempty: false
        description: "Marker ID from predefined list"
      name:
        type: string
        x-omitempty: false
      reference_range:
        $ref: '#/definitions/ReferenceRange'
      unit:
        $ref: '#/definitions/Unit'